+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/commands/CommandManager.java
@@ -32,20 +32,20 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 import net.minecraftforge.server.ServerLifecycleHooks;
 
-import de.markusbordihn.adaptiveperformancetweakscore.CoreConstants;
+import de.markusbordihn.adaptiveperformancetweakscore.Constants;
 
 @EventBusSubscriber
 public class CommandManager {
 
-  private static final Logger log = LogManager.getLogger(CoreConstants.LOG_NAME);
+  private static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
 
   protected CommandManager() {}
 
   @SubscribeEvent
   public static void handleRegisterCommandsEvent(RegisterCommandsEvent event) {
-    log.info("Registering /aptweaks commands for {} ...", CoreConstants.MOD_NAME);
+    log.info("Registering /aptweaks commands for {} ...", Constants.MOD_NAME);
     CommandDispatcher<CommandSourceStack> commandDispatcher = event.getDispatcher();
-    commandDispatcher.register(Commands.literal(CoreConstants.MOD_COMMAND)
+    commandDispatcher.register(Commands.literal(Constants.MOD_COMMAND)
     // @formatter:off
       .then(DebugCommand.register())
       .then(EntityCommand.register())