+++ items/src/main/java/de/markusbordihn/adaptiveperformancetweaksitems/Constants.java
@@ -21,13 +21,12 @@ package de.markusbordihn.adaptiveperformancetweaksitems;
 
 public final class Constants {
 
-    protected Constants() {
-    }
+    protected Constants() {}
 
     // General Mod definitions
-    public static final String ISSUE_REPORT = "https://github.com/MarkusBordihn/BOs-Adaptive-Performance-Tweaks/issues";
     public static final String LOG_NAME = "APTweaks(Items)";
     public static final String MOD_COMMAND = "aptweaks";
     public static final String MOD_ID = "adaptive_performance_tweaks_items";
     public static final String MOD_NAME = "Adaptive Performance Tweaks: Items";
+    public static final String MODULE_NAME = "Items";
 }