+++ player/src/main/java/de/markusbordihn/adaptiveperformancetweaksplayer/Constants.java
@@ -21,17 +21,12 @@ package de.markusbordihn.adaptiveperformancetweaksplayer;
 
 public final class Constants {
 
-    protected Constants() {
-    }
+    protected Constants() {}
 
     // General Mod definitions
-    public static final String ISSUE_REPORT = "https://github.com/MarkusBordihn/BOs-Adaptive-Performance-Tweaks/issues/new/choose";
-    public static final String LOG_NAME = "🚀 APTweaks(Player)";
+    public static final String LOG_NAME = "APTweaks(Player)";
     public static final String MOD_COMMAND = "aptweaks";
     public static final String MOD_ID = "adaptive_performance_tweaks_player";
-    public static final String MOD_NAME = "🚀 Adaptive Performance Tweaks: Player";
-
-    // Other
-    public static final String LOG_THANKS =
-            "🙏Thanks for using " + MOD_NAME + ", please report issues to " + ISSUE_REPORT;
+    public static final String MOD_NAME = "Adaptive Performance Tweaks: Player";
+    public static final String MODULE_NAME = "Player";
 }