+++ spawn/src/main/java/de/markusbordihn/adaptiveperformancetweaksspawn/Constants.java
@@ -24,15 +24,9 @@ public final class Constants {
   protected Constants() {}
 
   // General Mod definitions
-  public static final String ISSUE_REPORT =
-      "https://github.com/MarkusBordihn/BOs-Adaptive-Performance-Tweaks/issues/new/choose";
   public static final String LOG_NAME = "APTweaks(Spawn)";
   public static final String MOD_COMMAND = "aptweaks";
   public static final String MOD_ID = "adaptive_performance_tweaks_spawn";
   public static final String MOD_NAME = "Adaptive Performance Tweaks: Spawn";
-  public static final String MODULE_NAME = "Core";
-
-  // Other
-  public static final String LOG_THANKS =
-      "🙏Thanks for using " + MOD_NAME + ", please report issues to " + ISSUE_REPORT;
+  public static final String MODULE_NAME = "Spawn";
 }